File renaming and restructuring
