+++ code/player/Player.AFK.cs
