+++ code/player/Player.Confirmation.cs
