+++ code/player/Player.Flashlight.cs
